Add device CRUD
