+++ device-src/addDevice/event.json
