+++ device-src/completeDevice/event.json
