+++ device-src/deleteDevice/event.json
