+++ env.json
@@ -1,26 +1,26 @@
 {
-    "GetTodoFunction": {
-      "TABLE_NAME": "todo-table",
+    "GetDeviceFunction": {
+      "TABLE_NAME": "device-table",
       "AWS_NODEJS_CONNECTION_REUSE_ENABLED": "1"
     },
-    "GetAllTodoFunction": {
-      "TABLE_NAME": "todo-table",
+    "GetAllDeviceFunction": {
+      "TABLE_NAME": "device-table",
       "AWS_NODEJS_CONNECTION_REUSE_ENABLED": "1"
     },
-    "DeleteTodoFunction": {
-      "TABLE_NAME": "todo-table",
+    "DeleteDeviceFunction": {
+      "TABLE_NAME": "device-table",
       "AWS_NODEJS_CONNECTION_REUSE_ENABLED": "1"
     },
-    "CompleteTodoFunction": {
-      "TABLE_NAME": "todo-table",
+    "CompleteDeviceFunction": {
+      "TABLE_NAME": "device-table",
       "AWS_NODEJS_CONNECTION_REUSE_ENABLED": "1"
     },
-    "UpdateTodoFunction": {
-      "TABLE_NAME": "todo-table",
+    "UpdateDeviceFunction": {
+      "TABLE_NAME": "device-table",
       "AWS_NODEJS_CONNECTION_REUSE_ENABLED": "1"
     },
-    "AddTodoFunction": {
-      "TABLE_NAME": "todo-table",
+    "AddDeviceFunction": {
+      "TABLE_NAME": "device-table",
       "AWS_NODEJS_CONNECTION_REUSE_ENABLED": "1"
     }
 }
