--- todo-src/addTodo/package.json
@@ -1,14 +1,0 @@
-{
-  "name": "CreateTodo-item",
-  "version": "1.0.0",
-  "description": "Creates a ToDo item on the DynamoDB Table",
-  "main": "src/app.js",
-  "author": "SAM CLI",
-  "license": "MIT",
-  "dependencies": {
-    "aws-sdk": "^2.823.0",
-    "aws-xray-sdk-core": "^3.2.0",
-    "aws-embedded-metrics": "^2.0.2",
-    "uuid": "^8.3.2"
-  }
-}
