--- todo-src/test/environment/mac.json
@@ -1,28 +1,0 @@
-{
-    "GetTodoFunction": {
-        "ENDPOINT_OVERRIDE": "http://docker.for.mac.localhost:8000",
-        "TABLE_NAME": "TodoTable"
-    },
-    "GetAllTodoFunction": {
-        "ENDPOINT_OVERRIDE": "http://docker.for.mac.localhost:8000",
-        "TABLE_NAME": "TodoTable"
-    },
-    "CompleteTodoFunction": {
-        "ENDPOINT_OVERRIDE": "http://docker.for.mac.localhost:8000",
-        "TABLE_NAME": "TodoTable"
-    },
-    "AddTodoFunction": {
-        "ENDPOINT_OVERRIDE": "http://docker.for.mac.localhost:8000",
-        "TABLE_NAME": "TodoTable"
-    },
-    "UpdateTodoFunction": {
-        "ENDPOINT_OVERRIDE": "http://docker.for.mac.localhost:8000",
-        "TABLE_NAME": "TodoTable"
-    },
-    "DeleteTodoFunction": {
-        "ENDPOINT_OVERRIDE": "http://docker.for.mac.localhost:8000",
-        "TABLE_NAME": "TodoTable"
-    }
-
-    
-}
