+++ www/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "serverless-todo",
+  "name": "serverless-gamevision",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
