+++ www/public/index.html
@@ -6,7 +6,7 @@
   <link rel="shortcut icon" href="%PUBLIC_URL%/favicon.ico" />
   <meta name="viewport" content="width=device-width, initial-scale=1" />
   <meta name="theme-color" content="#000000" />
-  <meta name="description" content="Serverless Todo" />
+  <meta name="description" content="Serverless Gamevision" />
   <link rel="apple-touch-icon" href="logo192.png" />
   <link rel="stylesheet"
     href="https://cdnjs.cloudflare.com/ajax/libs/open-iconic/1.1.1/font/css/open-iconic-bootstrap.css"
@@ -20,7 +20,7 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
-  <title>Serverless Webapp on AWS</title>
+  <title>Gamevision</title>
 </head>
 
 <body>
