+++ www/src/config.default.js
@@ -3,7 +3,7 @@
 
 const config = {
   "aws_user_pools_web_client_id": "AWS_USER_POOLS_WEB_CLIENT_ID",     // CognitoClientID
-  "api_base_url": "API_BASE_URL",                                     // TodoFunctionApi
+  "api_base_url": "API_BASE_URL",                                     // DeviceFunctionApi
   "cognito_hosted_domain": "COGNITO_HOSTED_DOMAIN",                   // CognitoDomainName
   "redirect_url": "REDIRECT_URL"                                      // AmplifyURL
 };
