+++ www/src/config.js
@@ -3,7 +3,7 @@
 
 const config = {
   "aws_user_pools_web_client_id": "None",     // CognitoClientID
-  "api_base_url": "None",                                     // TodoFunctionApi
+  "api_base_url": "None",                                     //DeviceFunctionApi
   "cognito_hosted_domain": "None",                   // CognitoDomainName
   "redirect_url": "None"                                      // AmplifyURL
 };
